refact: shuffleAnswers
